requirements.txt
datasets
weaviate-client
langchain
langchain_google_genai
langchain-chroma
langchain-community
langchain-huggingface
langchain-cohere
langchain_experimental
langgraph
langsmith 
chroma
cohere
torch
transformers
sentence-transformers
huggingface-hub
jsonl2json
tdqm
python-dotenv